+++ retinazer-gwt-test/src/test/java/com/github/antag99/retinazer/GWTEntitySetListenerTest.java
@@ -77,8 +77,8 @@ public class GWTEntitySetListenerTest extends RetinazerTestCase {
         EntitySetListenerMock listenerB = new EntitySetListenerMock();
         EntitySetListenerMock listenerC = new EntitySetListenerMock();
         Engine engine = new Engine(new EngineConfig());
-        engine.getEntitiesFor(Family.with(FlagComponentB.class)).addListener(listenerB);
-        engine.getEntitiesFor(Family.with(FlagComponentC.class)).addListener(listenerC);
+        engine.getFamily(Family.with(FlagComponentB.class)).getEntities().addListener(listenerB);
+        engine.getFamily(Family.with(FlagComponentC.class)).getEntities().addListener(listenerC);
         Handle entity = engine.createEntity().cpy();
         engine.update();
         listenerB.verifyInserted(new int[0]);